requirements.txt
# Flask Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Database ORM
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23

# Database Migration
Flask-Migrate==4.0.5

# Form Handling and CSRF Protection
Flask-WTF==1.2.1
WTForms==3.1.1

# File Upload and Validation
Pillow==10.1.0

# AWS SDK
boto3==1.34.0
botocore==1.34.0

# Environment Variables
python-dotenv==1.0.0

# Development Tools
Flask-DebugToolbar==0.14.1

# Security
cryptography>=42.0.0

# Date and Time Utilities
python-dateutil==2.8.2

# HTTP Requests (for testing)
requests==2.31.0 